+++ sevenbridges/meta/transformer.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+
 import six
 from sevenbridges.errors import SbgError
 
@@ -6,7 +8,7 @@ class Transform(object):
     @staticmethod
     def to_project(project):
         from sevenbridges.models.project import Project
-        if project is None:
+        if not project:
             raise SbgError('Project is required!')
         elif isinstance(project, Project):
             return project.id
@@ -18,7 +20,7 @@ class Transform(object):
     @staticmethod
     def to_task(task):
         from sevenbridges.models.task import Task
-        if task is None:
+        if not task:
             raise SbgError('Task is required!')
         elif isinstance(task, Task):
             return task.id
@@ -30,7 +32,7 @@ class Transform(object):
     @staticmethod
     def to_app(app):
         from sevenbridges.models.app import App
-        if app is None:
+        if not app:
             raise SbgError('App is required!')
         elif isinstance(app, App):
             return app.id
@@ -42,7 +44,7 @@ class Transform(object):
     @staticmethod
     def to_file(file_):
         from sevenbridges.models.file import File
-        if file_ is None:
+        if not file_:
             raise SbgError('File is required!')
         elif isinstance(file_, File):
             return file_.id
@@ -54,7 +56,7 @@ class Transform(object):
     @staticmethod
     def to_user(user):
         from sevenbridges.models.user import User
-        if user is None:
+        if not user:
             raise SbgError('User is required!')
         elif isinstance(user, User):
             return user.username
@@ -66,7 +68,7 @@ class Transform(object):
     @staticmethod
     def to_billing_group(billing_group):
         from sevenbridges.models.billing_group import BillingGroup
-        if billing_group is None:
+        if not billing_group:
             raise SbgError('Billing group is required!')
         elif isinstance(billing_group, BillingGroup):
             return billing_group.id
@@ -78,7 +80,7 @@ class Transform(object):
     @staticmethod
     def to_volume(volume):
         from sevenbridges.models.volume import Volume
-        if volume is None:
+        if not volume:
             raise SbgError('Volume is required!')
         elif isinstance(volume, Volume):
             return volume.id
@@ -86,3 +88,12 @@ class Transform(object):
             return volume
         else:
             raise SbgError('Invalid volume parameter!')
+
+    @staticmethod
+    def to_datestring(d):
+        if not d:
+            raise SbgError('Date is required!')
+        elif isinstance(d, six.string_types):
+            return d
+        elif isinstance(d, datetime):
+            return d.isoformat().split('.', 1)[0]