+++ sevenbridges/models/storage_export.py
@@ -1,3 +1,4 @@
+import logging
 import six
 
 from sevenbridges.meta.resource import Resource
@@ -10,6 +11,8 @@ from sevenbridges.meta.fields import (
     HrefField, StringField, CompoundField, DateTimeField, BooleanField,
     DictField
 )
+
+log = logging.getLogger(__name__)
 
 
 # noinspection PyArgumentList
@@ -83,6 +86,11 @@ class Export(Resource):
         data['overwrite'] = overwrite
 
         api = api if api else cls._API
+        extra = {
+            'resource': cls.__name__,
+            'query': data
+        }
+        log.info('submit export', extra=extra)
         _export = api.post(cls._URL['query'], data=data).json()
         return Export(api=api, **_export)